+++ kernel/src/acpi/madt/ics/type1.rs
@@ -29,6 +29,6 @@ impl Type1 {
     }
 }
 
-pub fn handle(addr: u64) -> Result<(), crate::Error> {
+pub fn handle(addr: usize) -> Result<(), crate::Error> {
     Type1::get_ref(addr).handle()
 }
