+++ kernel/src/drivers/storage/nvme/command/admin/identify/active_namespace_id_list.rs
@@ -20,7 +20,7 @@ impl List {
     }
 }
 
-pub fn handle(addr: u64) -> Result<(), crate::Error> {
+pub fn handle(addr: usize) -> Result<(), crate::Error> {
     List::get_ref(addr).handle().delete()?;
     Ok(())
 }
