+++ kernel/src/drivers/storage/nvme/mod.rs
@@ -17,14 +17,14 @@ use queue::Queue;
 
 static mut NVME: NVMe = NVMe::null();
 
-pub fn set_config(addr: u64) {
+pub fn set_config(addr: usize) {
     unsafe { NVME.pcie_addr = addr };
 }
 
 struct NVMe {
-    pcie_addr: u64,
+    pcie_addr: usize,
 
-    addr: u64,
+    addr: usize,
 
     msi_x: MSIX,
 
@@ -64,21 +64,21 @@ impl NVMe {
     ///   - Bit 1: CRIMS for Controller Ready Independent of Media Support
     /// - Bit 61: NSSES for NVM Subsystem Shutdown Enhancements Supported
     /// - Bits 62 ..= 63: Reserved
-    const CAP: u64 = 0x0;
+    const CAP: usize = 0x0;
 
     /// Version
     /// - Bits 0 ..= 7: TER for Tertiary Version
     /// - Bits 8 ..= 15: MNR for Minor Version
     /// - Bits 16 ..= 31: MJR for Major Version
-    const VS: u64 = 0x8;
+    const VS: usize = 0x8;
 
     /// Interrupt Mask Set
     /// - Bits 0 ..= 31: IVMS for Interrupt Vector Mask Set
-    const INTMS: u64 = 0xC;
+    const INTMS: usize = 0xC;
 
     /// Interrupt Mask Clear
     /// - Bits 0 ..= 31: IVMC for Interrupt Vector Mask Clear
-    const INTMC: u64 = 0x10;
+    const INTMC: usize = 0x10;
 
     /// Controller Configuration
     /// - Bit 0: EN for Enable
@@ -109,7 +109,7 @@ impl NVMe {
     /// - Bits 20 ..= 23: IOCQES for I/O Completion Queue Entry Size
     /// - Bit 24: CRIME for Controller Ready Independent of Media Enable
     /// - Bits 25 ..= 31: Reserved
-    const CC: u64 = 0x14;
+    const CC: usize = 0x14;
 
     /// Controller Status
     /// - Bit 0: RDY for Ready
@@ -123,11 +123,11 @@ impl NVMe {
     /// - Bit 5: PP for Processing Paused
     /// - Bit 6: ST for Shutdown Type
     /// - Bit 7 ..= 31: Reserved
-    const CSTS: u64 = 0x1C;
+    const CSTS: usize = 0x1C;
 
     /// NVM Subsystem Reset
     /// - Bits 0 ..= 31: NSSRC for NVM Subsystem Reset Control
-    const NSSR: u64 = 0x20;
+    const NSSR: usize = 0x20;
 
     /// Admin Queue Attributes
     /// - Bits 0 ..= 11: ASQS for Admin Submission Queue Size
@@ -136,17 +136,17 @@ impl NVMe {
     /// - Bits 16 ..= 27: ACQS for Admin Completion Queue Size
     ///   - Valid range: 1 ..= 4095
     /// - Bits 28 ..= 31: Reserved
-    const AQA: u64 = 0x24;
+    const AQA: usize = 0x24;
 
     /// Admin Submission Queue
     /// - Bits 0 ..= 11: Reserved
     /// - Bits 12 ..= 63: ASQB for Admin Submission Queue Base
-    const ASQ: u64 = 0x28;
+    const ASQ: usize = 0x28;
 
     /// Admin Completion Queue
     /// - Bits 0 ..= 11: Reserved
     /// - Bits 12 ..= 63: ACQB for Admin Completion Queue Base
-    const ACQ: u64 = 0x30;
+    const ACQ: usize = 0x30;
 
     /// Controller Memory Buffer Location
     /// - Bits 0 ..= 2: BIR for Base Indicator Register
@@ -158,7 +158,7 @@ impl NVMe {
     /// - Bit 8: CQDA for CMB Queue Dword Alignment
     /// - Bits 9 ..= 11: Reserved
     /// - Bits 12 ..= 31: OFST for Offset
-    const CMBLOC: u64 = 0x38;
+    const CMBLOC: usize = 0x38;
 
     /// Controller Memory Buffer Size
     /// - Bit 0: SQS for Submission Queue Support
@@ -177,7 +177,7 @@ impl NVMe {
     ///   - 0x6: 64 GiB
     ///   - 0x7 ..= 0xF: Reserved
     /// - Bits 12 ..= 31: SZ for Size
-    const CMBSZ: u64 = 0x3C;
+    const CMBSZ: usize = 0x3C;
 
     /// Boot Partition Information
     /// - Bits 0 ..= 14: BPSZ for Boot Partition Size
@@ -189,31 +189,31 @@ impl NVMe {
     ///   - 0b11: Error completing Boot Partition read
     /// - Bits 26 ..= 30: Reserved
     /// - Bit 31: ABPID for Active Boot Partition ID
-    const BPINFO: u64 = 0x40;
+    const BPINFO: usize = 0x40;
 
     /// Boot Partition Read Select
     /// - Bits 0 ..= 9: BPRSZ for Boot Partition Read Size
     /// - Bits 10 ..= 29: BPROF for Boot Partition Read Offset
     /// - Bit 30: Reserved
     /// - Bit 31: BPID for Boot Partition Identifer
-    const BPRSEL: u64 = 0x44;
+    const BPRSEL: usize = 0x44;
 
     /// Boot Partition Memory Buffer Location
     /// - Bits 0 ..= 11: Reserved
     /// - Bits 12 ..= 63: BMBBA for Boot Partition Memory Buffer Base Address
-    const BPMBL: u64 = 0x48;
+    const BPMBL: usize = 0x48;
 
     /// Controller Memory Buffer Memory Space Control
     /// - Bit 0: CRE for Capabilities Registers Enabled
     /// - Bit 1: CMSE for Controller Memory Space Enable
     /// - Bits 2 ..= 11: Reserved
     /// - Bits 12 ..= 63: CBA for Controller Base Address
-    const CMBMSC: u64 = 0x50;
+    const CMBMSC: usize = 0x50;
 
     /// Controller Memory Buffer Status
     /// - Bit 0: CBAI for Controller Base Address Invalid
     /// - Bits 1 ..= 31: Reserved
-    const CMBSTS: u64 = 0x58;
+    const CMBSTS: usize = 0x58;
 
     /// Controller Memory Buffer Elasticity Buffer Size
     /// - Bits 0 ..= 3: CMBSZU for CMB Elasticity Buffer Size Units
@@ -225,7 +225,7 @@ impl NVMe {
     /// - Bit 4: CMBRBB for CMB Read Bypass Behavior
     /// - Bits 5 ..= 7: Reserved
     /// - Bits 8 ..= 31: CMBWBZ for CMB Elasticity Buffer Size Base
-    const CMBEBS: u64 = 0x5C;
+    const CMBEBS: usize = 0x5C;
 
     /// Controller Memory Buffer Sustained Write Throughput
     /// - Bits 0 ..= 3: CMBSWTU for CMB Sustained Write Throughput Units
@@ -236,16 +236,16 @@ impl NVMe {
     ///   - 0x4 ..= 0xF: Reserved
     /// - Bits 4 ..= 7: Reserved
     /// - Bits 8 ..= 31: CMBSWTV for CMB Sustained Write Throughput
-    const CMBSWTP: u64 = 0x60;
+    const CMBSWTP: usize = 0x60;
 
     /// NVM Subsystem Shutdown
     /// - Bits 0 ..= 31: NSSC for NVM Subsystem Shutdown Control
-    const NSSD: u64 = 0x64;
+    const NSSD: usize = 0x64;
 
     /// Controller Ready Timeouts
     /// - Bits 0 ..= 15: CRWMT for Controller Ready With Media Timeout
     /// - Bits 16 ..= 31: CRIMT for Controller Ready Independent of Media Timeout
-    const CRTO: u64 = 0x68;
+    const CRTO: usize = 0x68;
 
     /// Persistent Memory Capabilities
     /// - Bits 0 ..= 2: Reserved
@@ -264,12 +264,12 @@ impl NVMe {
     /// - Bits 16 ..= 23: PMRTO for Persistent Memory Region Timeout
     /// - Bit 24: CMSS for Controller Memory Space Supported
     /// - Bits 25 ..= 31: Reserved
-    const PMRCAP: u64 = 0xE00;
+    const PMRCAP: usize = 0xE00;
 
     /// Persistent Memory Region Control
     /// - Bit 0: EN for Enable
     /// - Bits 1 ..= 31: Reserved
-    const PMRCTL: u64 = 0xE04;
+    const PMRCTL: usize = 0xE04;
 
     /// Persistent Memory Region Status
     /// - Bits 0 ..= 7: ERR for Error
@@ -282,7 +282,7 @@ impl NVMe {
     ///   - 0b100 ..= 0b111: Reserved
     /// - Bit 12: CBAI for Controller Base Address Invalid
     /// - Bits 13 ..= 31: Reserved
-    const PMRSTS: u64 = 0xE08;
+    const PMRSTS: usize = 0xE08;
 
     /// Persistent Memory Region Elasticity Buffer Size
     /// - Bits 0 ..= 3: PMRSZU for PMR Elasticity Buffer Size Units
@@ -294,28 +294,28 @@ impl NVMe {
     /// - Bit 4: PMRRBB for PMR Read Bypass Behavior
     /// - Bits 5 ..= 7: Reserved
     /// - Bits 8 ..= 31: PMRWBZ for PMR Elasticity Buffer Size Base
-    const PMREBS: u64 = 0xE0C;
+    const PMREBS: usize = 0xE0C;
 
     /// Persistent Memory Region Sustained Write Throughput
     /// - Bits 0 ..= 3: PMRSWTU for PMR Sustained Write Throughput Units
     /// - Bits 4 ..= 7: Reserved
     /// - Bits 8 ..= 31: PMRSWTV for PMR Sustained Write Throughput
-    const PMRSWTP: u64 = 0xE10;
+    const PMRSWTP: usize = 0xE10;
 
     /// Persistent Memory Region Controller Memory Space Control Lower
     /// - Bit 0: Reserved
     /// - Bit 1: CMSE for Controller Memory Space Enable
     /// - Bits 2 ..= 11: Reserved
     /// - Bits 12 ..= 31: CBA for Controller Base Address
-    const PMRMSCL: u64 = 0xE14;
+    const PMRMSCL: usize = 0xE14;
 
     /// Persistent Memory Region Controller Memory Space Control Upper
     /// - Bits 0 ..= 31: CBA for Controller Base Address
-    const PMRMSCU: u64 = 0xE18;
+    const PMRMSCU: usize = 0xE18;
 
     /// - Submission: Base + (2 * X) * (1 << (2 + CAP.DSTRD))
     /// - Completion: Base + (2 * X + 1) * (1 << (2 + CAP.DSTRD))
-    const DOORBELL: u64 = 0x1000;
+    const DOORBELL: usize = 0x1000;
 
     const fn null() -> Self {
         Self {
@@ -327,14 +327,14 @@ impl NVMe {
         }
     }
 
-    fn is_u64_register(offset: u64) -> bool {
+    fn is_u64_register(offset: usize) -> bool {
         matches!(
             offset,
             Self::CAP | Self::ASQ | Self::ACQ | Self::BPMBL | Self::CMBMSC
         )
     }
 
-    fn read(&self, offset: u64) -> u64 {
+    fn read(&self, offset: usize) -> u64 {
         unsafe {
             if Self::is_u64_register(offset) {
                 ((self.addr + offset) as *const u64).read_volatile()
@@ -344,7 +344,7 @@ impl NVMe {
         }
     }
 
-    fn write(&self, offset: u64, val: u64) {
+    fn write(&self, offset: usize, val: u64) {
         unsafe {
             if Self::is_u64_register(offset) {
                 ((self.addr + offset) as *mut u64).write_volatile(val)
@@ -399,8 +399,8 @@ impl NVMe {
             Self::AQA,
             ((admin_size as u64 - 1) << 16) | (admin_size as u64 - 1),
         );
-        self.write(Self::ASQ, asq);
-        self.write(Self::ACQ, acq);
+        self.write(Self::ASQ, asq as u64);
+        self.write(Self::ACQ, acq as u64);
         self.write(Self::CC, {
             let cap = self.read(Self::CAP);
             1 | ({
