+++ kernel/src/io/mod.rs
@@ -4,7 +4,7 @@ pub mod port;
 pub mod text;
 
 pub fn init(
-    frame_buffer_base: u64,
+    frame_buffer_base: usize,
     screen_width: usize,
     screen_height: usize,
     screen_stride: usize,
