+++ kernel/src/io/text/output/mod.rs
@@ -106,7 +106,7 @@ impl Output for &str {
 }
 
 pub fn init(
-    frame_buffer_base: u64,
+    frame_buffer_base: usize,
     screen_width: usize,
     screen_height: usize,
     screen_stride: usize,
